// src/app/shared/enums/FiledTypes.ts
export enum FieldType{
  NONE = "None",
  CHECKBOX = 'Checkbox',
  TEXT = 'Text',
  NUMERIC = 'Numeric',
  TEXTAREA = 'Text Area',
  RADIO = 'Radio',
  SELECTION = 'Selection',
  MULTIPLE_SELECTION = 'Multiple Selection',
  LIMITED_MULTIPLE_SELECTION = 'Limited Multiple Selection',
  SPECIAL = 'Special',
}